Split the Alerting List Page out into a separate component
