+++ src/components/alerting/AlertUtils.tsx
@@ -1,0 +1,199 @@
+import * as React from 'react';
+import {
+  Action,
+  Alert,
+  AlertSeverity,
+  AlertStates,
+  BlueInfoCircleIcon,
+  PrometheusAlert,
+  RedExclamationCircleIcon,
+  Rule,
+  Timestamp,
+  YellowExclamationTriangleIcon,
+} from '@openshift-console/dynamic-plugin-sdk';
+import { AlertSource, MonitoringResource } from '../types';
+import * as _ from 'lodash-es';
+import { useTranslation } from 'react-i18next';
+import { Alert as PFAlert } from '@patternfly/react-core';
+import { labelsToParams, RuleResource, SilenceResource } from '../utils';
+import classNames from 'classnames';
+import { BellIcon, BellSlashIcon, OutlinedBellIcon } from '@patternfly/react-icons';
+
+export const getAdditionalSources = <T extends Alert | Rule>(
+  data: Array<T>,
+  itemSource: (item: T) => string,
+) => {
+  if (data) {
+    const additionalSources = new Set<string>();
+    data.forEach((item) => {
+      const source = itemSource(item);
+      if (source !== AlertSource.Platform && source !== AlertSource.User) {
+        additionalSources.add(source);
+      }
+    });
+    return Array.from(additionalSources).map((item) => ({ id: item, title: _.startCase(item) }));
+  }
+  return [];
+};
+
+export const alertingRuleSource = (rule: Rule): AlertSource | string => {
+  if (rule.sourceId === undefined || rule.sourceId === 'prometheus') {
+    return rule.labels?.prometheus === 'openshift-monitoring/k8s'
+      ? AlertSource.Platform
+      : AlertSource.User;
+  }
+
+  return rule.sourceId;
+};
+
+export const alertSource = (alert: Alert): AlertSource | string => alertingRuleSource(alert.rule);
+
+export const SilencesNotLoadedWarning: React.FC<{ silencesLoadError: any }> = ({
+  silencesLoadError,
+}) => {
+  const { t } = useTranslation('plugin__monitoring-plugin');
+
+  return (
+    <PFAlert
+      className="co-alert"
+      isInline
+      title={t(
+        'Error loading silences from Alertmanager. Some of the alerts below may actually be silenced.',
+      )}
+      variant="warning"
+    >
+      {silencesLoadError.json?.error || silencesLoadError.message}
+    </PFAlert>
+  );
+};
+
+type ActionWithHref = Omit<Action, 'cta'> & { cta: { href: string; external?: boolean } };
+export const isActionWithHref = (action: Action): action is ActionWithHref => 'href' in action.cta;
+
+export const ruleURL = (rule: Rule) => `${RuleResource.plural}/${_.get(rule, 'id')}`;
+
+export const silenceAlertURL = (alert: PrometheusAlert) =>
+  `${SilenceResource.plural}/~new?${labelsToParams(alert.labels)}`;
+
+type ActionWithCallBack = Omit<Action, 'cta'> & { cta: () => void };
+export const isActionWithCallback = (action: Action): action is ActionWithCallBack =>
+  typeof action.cta === 'function';
+
+export const MonitoringResourceIcon: React.FC<MonitoringResourceIconProps> = ({
+  className,
+  resource,
+}) => (
+  <span
+    className={classNames(
+      `co-m-resource-icon co-m-resource-${resource.kind.toLowerCase()}`,
+      className,
+    )}
+    title={resource.label}
+  >
+    {resource.abbr}
+  </span>
+);
+
+type MonitoringResourceIconProps = {
+  className?: string;
+  resource: MonitoringResource;
+};
+
+export const Severity: React.FC<{ severity: string }> = React.memo(({ severity }) => {
+  const { t } = useTranslation('plugin__monitoring-plugin');
+
+  const getSeverityKey = (severityData: string) => {
+    switch (severityData) {
+      case AlertSeverity.Critical:
+        return t('Critical');
+      case AlertSeverity.Info:
+        return t('Info');
+      case AlertSeverity.Warning:
+        return t('Warning');
+      case AlertSeverity.None:
+        return t('None');
+      default:
+        return severityData;
+    }
+  };
+
+  return _.isNil(severity) ? (
+    <>-</>
+  ) : (
+    <>
+      <SeverityIcon severity={severity} /> {getSeverityKey(severity)}
+    </>
+  );
+});
+
+export const SeverityIcon: React.FC<{ severity: string }> = React.memo(({ severity }) => {
+  const Icon =
+    {
+      [AlertSeverity.Critical]: RedExclamationCircleIcon,
+      [AlertSeverity.Info]: BlueInfoCircleIcon,
+      [AlertSeverity.None]: BlueInfoCircleIcon,
+      [AlertSeverity.Warning]: YellowExclamationTriangleIcon,
+    }[severity] || YellowExclamationTriangleIcon;
+  return <Icon />;
+});
+
+export const AlertState: React.FC<AlertStateProps> = React.memo(({ state }) => {
+  const { t } = useTranslation('plugin__monitoring-plugin');
+
+  const icon = <AlertStateIcon state={state} />;
+
+  return icon ? (
+    <>
+      {icon} {getAlertStateKey(state, t)}
+    </>
+  ) : null;
+});
+
+type AlertStateProps = {
+  state: AlertStates;
+};
+
+export const AlertStateIcon: React.FC<{ state: string }> = React.memo(({ state }) => {
+  switch (state) {
+    case AlertStates.Firing:
+      return <BellIcon />;
+    case AlertStates.Pending:
+      return <OutlinedBellIcon />;
+    case AlertStates.Silenced:
+      return <BellSlashIcon className="text-muted" />;
+    default:
+      return null;
+  }
+});
+
+export const getAlertStateKey = (state, t) => {
+  switch (state) {
+    case AlertStates.Firing:
+      return t('Firing');
+    case AlertStates.Pending:
+      return t('Pending');
+    case AlertStates.Silenced:
+      return t('Silenced');
+    default:
+      return t('Not Firing');
+  }
+};
+
+export const AlertStateDescription: React.FC<{ alert: Alert }> = ({ alert }) => {
+  const { t } = useTranslation('plugin__monitoring-plugin');
+
+  if (alert && !_.isEmpty(alert.silencedBy)) {
+    return <StateTimestamp text={t('Ends')} timestamp={_.max(_.map(alert.silencedBy, 'endsAt'))} />;
+  }
+  if (alert && alert.activeAt) {
+    return <StateTimestamp text={t('Since')} timestamp={alert.activeAt} />;
+  }
+  return null;
+};
+
+const StateTimestamp = ({ text, timestamp }) => (
+  <div className="text-muted monitoring-timestamp">
+    {text}&nbsp;
+    <Timestamp timestamp={timestamp} />
+  </div>
+);
